Loosen type bounds
